+++ splashdamage/idlib/hashing/CRC8.cpp
@@ -22,7 +22,7 @@ static byte crctable[256];
 
 void make_crc_table( void ) {
 	int i, j;
-	unsigned long poly, c;
+	unsigned /* 64long */int poly, c;
 	/* terms of polynomial defining this crc (except x^8): */
 	static const byte p[] = {0,1,2};
